Move edit sessions services into contrib since they are not accessed from outside the contrib
