+++ src/vs/workbench/contrib/sessionSync/common/sessionSync.ts
